fix(desktop): align status bar context overflow
Problem:
The status bar still clamped over-limit context usage to 100%, so it disagreed with the context panel and ring.

Root cause:
The status bar kept a local percentage calculation instead of using the shared raw/display context-window contract.

Fix:
Move contextWindowPercentages into a shared frontend library and use its raw percentage for the status bar label while preserving capped graphical fills.

Verification:
- 114 focused context and status-bar assertions
- pnpm build
- Playwright Chromium DOM, visibility, and console check
